Rename map file to geolocation
