+++ frontend/src/interfaces/geolocation.ts
