+++ frontend/src/interfaces/index.ts
@@ -1,1 +1,1 @@
-export * from "./map";
+export * from "./geolocation";
